cast as float32

+++ ctr_generic_envs/envs/exact_model.py
@@ -50,7 +50,7 @@ class ExactModel(object):
         uz_0_ = np.array([0, 0, 0])
         shape, U_z, tip = self.ctr_model(uz_0_, alpha_0_, r_0_, R_0_, segment, beta)
         assert not np.any(np.isnan(shape))
-        return shape[-1]
+        return (shape[-1]).astype(np.float32)
         #return np.random.rand(3)
 
     def get_r(self):
